Update README.md
This is an update for the forked repository to ensure the open-source contribution remains open-source and any credit is properly and professionally given for included works. I created the Ender Controller that connects to my 3D printer via USB, then I found a serverless "server" that I could use to act as my peer to peer connection, a.k.a the Peer to Peer Cue application I integrated into my controller for fully remote controlling of the printer from anywhere, without the need for additional code.